refactor(extract): rename the extraction source files to mirror the endpoints
parse.ts implements POST /extraction/parse and extract.ts implements POST
/extraction/extract, so name the files (and their tests) for the endpoints they
serve. Pure rename plus import-path updates; no behavior change.
