+++ tests/extract.test.ts
@@ -4,50 +4,46 @@ import os from 'os'
 import path from 'path'
 import { Readable } from 'stream'
 import { setSandboxDirectory } from '../src/fs/sandbox.js'
-import { performParseDocumentCall, SAME_PATH_ERROR } from '../src/dws/extract.js'
+import { performExtractFieldsCall } from '../src/dws/extract.js'
+import { SAME_PATH_ERROR } from '../src/dws/parse.js'
 import type { DwsApiClient } from '../src/dws/client.js'
-import type { ParseDocumentArgs } from '../src/schemas.js'
+import { ExtractFieldsArgsSchema } from '../src/schemas.js'
+import type { ExtractFieldsArgs } from '../src/schemas.js'
 import type { CallToolResult } from '@modelcontextprotocol/sdk/types.js'
 
-// A recognizable "PII" string used to prove extracted content never appears in
-// the inline spatial summary (it must only live in the written file).
+// A recognizable "PII" value used to prove citation content only lands in the
+// written file, never inline (the schema's field NAMES are always inline).
 const SECRET = 'SSN 123-45-6789'
 
-const spatialFixture = {
+const invoiceSchema = {
+  type: 'object' as const,
+  properties: {
+    invoiceNumber: { type: 'string' },
+    total: { type: 'number' },
+  },
+  required: ['invoiceNumber'],
+}
+
+const extractFixture = {
   status: 200,
   requestId: 'req_test',
+  runId: 'run_abc123',
   output: {
-    elements: [
-      {
-        id: '1',
-        type: 'paragraph',
-        role: 'Title',
-        text: 'Quarterly Report',
-        confidence: 0.95,
-        readingOrder: 0,
-        bounds: { x: 100, y: 50, width: 400, height: 35 },
-        page: { pageIndex: 0, pageNumber: 1, width: 1818, height: 2422 },
-      },
-      {
-        id: '2',
-        type: 'keyValueRegion',
-        text: SECRET,
-        confidence: 0.4,
-        readingOrder: 1,
-        bounds: { x: 100, y: 200, width: 300, height: 20 },
-        page: { pageIndex: 0, pageNumber: 1, width: 1818, height: 2422 },
-      },
-      {
-        id: '3',
-        type: 'table',
-        confidence: 0.8,
-        readingOrder: 2,
-        bounds: { x: 100, y: 400, width: 600, height: 300 },
-        page: { pageIndex: 1, pageNumber: 2, width: 1818, height: 2422 },
-      },
-    ],
+    data: { invoiceNumber: 'INV-001', total: 42.5 },
+    metadata: {
+      invoiceNumber: { bbox: [1, 2, 3, 4], confidence: 0.95, match: 'id_match' },
+      total: { bbox: [5, 6, 7, 8], confidence: 0.4, match: 'fuzzy_match', note: SECRET },
+    },
+    pages: [{ pageIndex: 0 }],
   },
-  metrics: { processingTimeMs: 100, pagesProcessed: 2 },
+  metrics: { pagesProcessed: 1 },
+  usage: {
+    data_extraction_credits: { cost: 10.5, remainingCredits: 989.5 },
+    price_composition: {
+      parse: { units: 1, unit_cost: 9, cost: 9, currency: 'credits' },
+      extract: { units: 1, unit_cost: 1.5, cost: 1.5, currency: 'credits' },
+    },
+  },
 }
 
 function mockClient(payload: unknown): { client: DwsApiClient; post: ReturnType<typeof vi.fn> } {
@@ -55,7 +51,6 @@ function mockClient(payload: unknown): { client: DwsApiClient; post: ReturnType<
   return { client: { post } as unknown as DwsApiClient, post }
 }
 
-/** A client whose POST rejects with an axios-shaped error carrying a streamed body. */
 function mockErrorClient(
   status: number,
   payload: unknown,
@@ -70,7 +65,6 @@ function mockErrorClient(
   return { client: { post } as unknown as DwsApiClient, post }
 }
 
-/** Parses the `instructions` field sent in a multipart form-data payload. */
 function parseFormInstructions(form: { getBuffer: () => Buffer }): Record<string, unknown> {
   const raw = form.getBuffer().toString('utf-8')
   const match = raw.match(/name="instructions"\r\n\r\n([\s\S]*?)\r\n--/)
@@ -103,260 +97,60 @@ async function writeInput(): Promise<string> {
   return name
 }
 
-function extractArgs(overrides: Partial<ParseDocumentArgs>): ParseDocumentArgs {
+function extractArgs(overrides: Partial<ExtractFieldsArgs>): ExtractFieldsArgs {
   const noDocumentGiven = !('filePath' in overrides) && !('url' in overrides)
   return {
     filePath: noDocumentGiven ? `input-${counter}.pdf` : overrides.filePath,
     url: overrides.url,
+    schema: overrides.schema ?? invoiceSchema,
+    instructions: overrides.instructions,
     mode: overrides.mode ?? 'understand',
-    format: overrides.format,
-    formats: overrides.formats,
-    includeWords: overrides.includeWords,
     language: overrides.language,
     maxLanguages: overrides.maxLanguages,
     maxScripts: overrides.maxScripts,
-    useHtmlTables: overrides.useHtmlTables,
-    enableSemanticBlockFormatting: overrides.enableSemanticBlockFormatting,
-    includeHeadersAndFooters: overrides.includeHeadersAndFooters,
-    extractWordsFromPictures: overrides.extractWordsFromPictures,
+    includeCitations: overrides.includeCitations,
+    strict: overrides.strict,
+    multimodal: overrides.multimodal,
     storeRun: overrides.storeRun ?? false,
     outputPath: overrides.outputPath,
   }
 }
 
-describe('performParseDocumentCall', () => {
-  it('returns markdown output inline', async () => {
-    const input = await writeInput()
-    const { client, post } = mockClient({ output: { markdown: '# Hello World' } })
-
-    const result = await performParseDocumentCall(extractArgs({ filePath: input, mode: 'text', format: 'markdown' }), client)
-
-    expect(result.isError).toBeFalsy()
-    expect(text(result)).toBe('# Hello World')
-    expect(post).toHaveBeenCalledOnce()
-  })
-
-  it('writes markdown to a file when outputPath is given, returning a summary not the content', async () => {
-    const input = await writeInput()
-    const outName = `out-${counter}.md`
-    const { client } = mockClient({ output: { markdown: '# Big Document\n\nlots of text' } })
-
-    const result = await performParseDocumentCall(
-      extractArgs({ filePath: input, mode: 'text', format: 'markdown', outputPath: outName }),
-      client,
-    )
-
-    expect(result.isError).toBeFalsy()
-    const summary = text(result)
-    expect(summary).toContain('Wrote')
-    expect(summary).toContain(outName)
-    expect(summary).not.toContain('lots of text')
-    const written = await fs.promises.readFile(path.join(sandboxDir, outName), 'utf-8')
-    expect(written).toBe('# Big Document\n\nlots of text')
-  })
-
-  it('rejects a 2xx response with no spatial element list without writing the file', async () => {
-    const input = await writeInput()
-    const outName = `out-${counter}.json`
-    const { client } = mockClient({ status: 200, output: { markdown: 'oops wrong shape' } })
-
-    const result = await performParseDocumentCall(
-      extractArgs({ filePath: input, mode: 'structure', format: 'spatial', outputPath: outName }),
-      client,
-    )
-
-    expect(result.isError).toBe(true)
-    expect(text(result)).toContain('output.elements')
-    await expect(fs.promises.access(path.join(sandboxDir, outName))).rejects.toThrow()
-  })
-
-  it('writes spatial output to a file and returns a content-free summary', async () => {
-    const input = await writeInput()
-    const outName = `out-${counter}.json`
-    const { client } = mockClient(spatialFixture)
-
-    const result = await performParseDocumentCall(
-      extractArgs({ filePath: input, mode: 'structure', format: 'spatial', outputPath: outName }),
-      client,
-    )
-
-    expect(result.isError).toBeFalsy()
-    const summary = text(result)
-    // Summary reports structure, not content.
-    expect(summary).toContain('Extracted 3 elements')
-    expect(summary).toContain('keyValueRegion: 1')
-    expect(summary).toContain('Low-confidence elements')
-    // The PII must NOT leak into the inline summary...
-    expect(summary).not.toContain(SECRET)
-    // ...but the full data IS persisted to the file.
-    const written = await fs.promises.readFile(path.join(sandboxDir, outName), 'utf-8')
-    expect(written).toContain(SECRET)
-  })
-
-  it('rejects spatial output without an outputPath, before any API call', async () => {
-    const input = await writeInput()
-    const { client, post } = mockClient(spatialFixture)
-
-    const result = await performParseDocumentCall(
-      extractArgs({ filePath: input, mode: 'structure', format: 'spatial' }),
-      client,
-    )
-
-    expect(result.isError).toBe(true)
-    expect(text(result)).toContain('outputPath')
-    expect(post).not.toHaveBeenCalled()
-  })
-
-  it('rejects text mode with spatial output', async () => {
-    const input = await writeInput()
-    const { client, post } = mockClient(spatialFixture)
-
-    const result = await performParseDocumentCall(
-      extractArgs({ filePath: input, mode: 'text', format: 'spatial', outputPath: `out-${counter}.json` }),
-      client,
-    )
-
-    expect(result.isError).toBe(true)
-    expect(text(result)).toContain('text mode')
-    expect(post).not.toHaveBeenCalled()
-  })
-
-  it('contains an outside-sandbox absolute outputPath within the sandbox', async () => {
-    const input = await writeInput()
-    const escape = path.join(os.tmpdir(), `escape-${counter}.json`)
-    const { client } = mockClient(spatialFixture)
-
-    const result = await performParseDocumentCall(
-      extractArgs({ filePath: input, mode: 'structure', format: 'spatial', outputPath: escape }),
-      client,
-    )
-
-    // The sandbox re-roots the absolute path inside the sandbox rather than
-    // writing to the literal location, so nothing escapes.
-    expect(result.isError).toBeFalsy()
-    await expect(fs.promises.access(escape)).rejects.toThrow()
-  })
-
-  it('sends the x-nutrient-api-version header on the extraction call', async () => {
-    const input = await writeInput()
-    const { client, post } = mockClient({ output: { markdown: '# Hello' } })
-
-    await performParseDocumentCall(extractArgs({ filePath: input, mode: 'text', format: 'markdown' }), client)
-
-    expect(post).toHaveBeenCalledWith(expect.any(String), expect.anything(), { 'x-nutrient-api-version': '2026-05-25' })
-  })
-
-  it('surfaces Data Extraction credit usage in the success message, distinct from Processor credits', async () => {
-    const input = await writeInput()
-    const { client } = mockClient({
-      output: { markdown: '# Hello' },
-      usage: { data_extraction_credits: { cost: 9, remainingCredits: 991 } },
-    })
-
-    const result = await performParseDocumentCall(extractArgs({ filePath: input, mode: 'text', format: 'markdown' }), client)
-
-    const summary = text(result)
-    expect(summary).toContain('9 Data Extraction credit')
-    expect(summary).toContain('991 remaining')
-    expect(summary).toContain('separate')
-  })
-
-  it('surfaces runId in the success message when storeRun is true', async () => {
-    const input = await writeInput()
-    const { client } = mockClient({ output: { markdown: '# Hello' }, runId: 'run_abc123' })
-
-    const result = await performParseDocumentCall(
-      extractArgs({ filePath: input, mode: 'text', format: 'markdown', storeRun: true }),
-      client,
-    )
-
-    expect(text(result)).toContain('run_abc123')
-  })
-
-  it('reports remaining Data Extraction credits even when the response omits cost', async () => {
-    const input = await writeInput()
-    const { client } = mockClient({
-      output: { markdown: '# Hello' },
-      usage: { data_extraction_credits: { remainingCredits: 3 } },
-    })
-
-    const result = await performParseDocumentCall(extractArgs({ filePath: input, mode: 'text', format: 'markdown' }), client)
-
-    const summary = text(result)
-    expect(summary).toContain('3 remaining')
-    expect(summary).toContain('separate')
-  })
-
-  it('rejects an outputPath that resolves to the same file as filePath, before any API call', async () => {
-    const input = await writeInput()
-    const { client, post } = mockClient(spatialFixture)
-
-    const result = await performParseDocumentCall(
-      extractArgs({ filePath: input, mode: 'structure', format: 'spatial', outputPath: input }),
-      client,
-    )
-
-    expect(result.isError).toBe(true)
-    expect(text(result)).toBe(SAME_PATH_ERROR)
-    expect(post).not.toHaveBeenCalled()
-  })
-
-  it('reports a post-billing write failure as billed rather than as an API error', async () => {
-    const input = await writeInput()
-    const outName = `out-${counter}.json`
-    const { client } = mockClient(spatialFixture)
-    const writeFileSpy = vi.spyOn(fs.promises, 'writeFile').mockRejectedValueOnce(new Error('ENOSPC: no space left'))
-
-    const result = await performParseDocumentCall(
-      extractArgs({ filePath: input, mode: 'structure', format: 'spatial', outputPath: outName }),
-      client,
-    )
-    writeFileSpy.mockRestore()
-
-    expect(result.isError).toBe(true)
-    const message = text(result)
-    expect(message).toContain('succeeded and was billed')
-    expect(message).toContain('billed again')
-    expect(message).not.toContain('Data Extraction API error')
-  })
-
+describe('performExtractFieldsCall', () => {
   describe('document input: filePath vs url', () => {
     it('sends a multipart request with a JSON-stringified instructions field for filePath', async () => {
       const input = await writeInput()
-      const { client, post } = mockClient({ output: { markdown: '# Hello' } })
+      const { client, post } = mockClient(extractFixture)
 
-      await performParseDocumentCall(extractArgs({ filePath: input, mode: 'text', format: 'markdown' }), client)
+      const result = await performExtractFieldsCall(extractArgs({ filePath: input }), client)
 
+      expect(result.isError).toBeFalsy()
       const form = post.mock.calls[0][1]
       expect(form.constructor.name).toBe('FormData')
       const instructions = parseFormInstructions(form)
-      expect(instructions).toMatchObject({ mode: 'text', output: { format: 'markdown' } })
+      expect(instructions).toMatchObject({ schema: invoiceSchema, parseConfig: { mode: 'understand' } })
     })
 
-    it('sends a JSON body containing the url for url input, not multipart', async () => {
-      const { client, post } = mockClient({ output: { markdown: '# Hello' } })
+    it('sends a JSON body containing the url and schema for url input, not multipart', async () => {
+      const { client, post } = mockClient(extractFixture)
 
-      const result = await performParseDocumentCall(
-        extractArgs({ filePath: undefined, url: 'https://example.com/doc.pdf', mode: 'text', format: 'markdown' }),
+      const result = await performExtractFieldsCall(
+        extractArgs({ filePath: undefined, url: 'https://example.com/doc.pdf' }),
         client,
       )
 
       expect(result.isError).toBeFalsy()
       const body = post.mock.calls[0][1]
-      expect(body).toMatchObject({
-        mode: 'text',
-        output: { format: 'markdown' },
-        url: 'https://example.com/doc.pdf',
-      })
+      expect(body).toMatchObject({ schema: invoiceSchema, url: 'https://example.com/doc.pdf' })
       expect(body.constructor.name).not.toBe('FormData')
     })
 
-    it('rejects when both filePath and url are provided', async () => {
+    it('rejects when both filePath and url are provided, before any API call', async () => {
       const input = await writeInput()
-      const { client, post } = mockClient({ output: { markdown: '# Hello' } })
+      const { client, post } = mockClient(extractFixture)
 
-      const result = await performParseDocumentCall(
-        extractArgs({ filePath: input, url: 'https://example.com/doc.pdf', mode: 'text', format: 'markdown' }),
+      const result = await performExtractFieldsCall(
+        extractArgs({ filePath: input, url: 'https://example.com/doc.pdf' }),
         client,
       )
 
@@ -365,261 +159,68 @@ describe('performParseDocumentCall', () => {
       expect(post).not.toHaveBeenCalled()
     })
 
-    it('rejects when neither filePath nor url is provided', async () => {
-      const { client, post } = mockClient({ output: { markdown: '# Hello' } })
+    it('rejects when neither filePath nor url is provided, before any API call', async () => {
+      const { client, post } = mockClient(extractFixture)
 
-      const result = await performParseDocumentCall(
-        extractArgs({ filePath: undefined, url: undefined, mode: 'text', format: 'markdown' }),
-        client,
-      )
+      const result = await performExtractFieldsCall(extractArgs({ filePath: undefined, url: undefined }), client)
 
       expect(result.isError).toBe(true)
       expect(text(result)).toContain('exactly one')
       expect(post).not.toHaveBeenCalled()
     })
-  })
 
-  describe('format vs formats', () => {
-    it('rejects format and formats given together, before any API call', async () => {
+    it('rejects an outputPath that resolves to the same file as filePath, before any API call', async () => {
       const input = await writeInput()
-      const { client, post } = mockClient(spatialFixture)
+      const { client, post } = mockClient(extractFixture)
 
-      const result = await performParseDocumentCall(
-        extractArgs({
-          filePath: input,
-          format: 'markdown',
-          formats: ['spatial', 'markdown'],
-          outputPath: `out-${counter}.json`,
-        }),
-        client,
-      )
+      const result = await performExtractFieldsCall(extractArgs({ filePath: input, outputPath: input }), client)
 
       expect(result.isError).toBe(true)
-      expect(text(result)).toContain('one of format or formats')
+      expect(text(result)).toBe(SAME_PATH_ERROR)
       expect(post).not.toHaveBeenCalled()
-    })
-
-    it('sends output.formats (and never output.format) for a multi-format request', async () => {
-      const input = await writeInput()
-      const markdown = '# Quarterly Report'
-      const { client, post } = mockClient({ ...spatialFixture, output: { ...spatialFixture.output, markdown } })
-
-      await performParseDocumentCall(
-        extractArgs({
-          filePath: input,
-          mode: 'structure',
-          formats: ['spatial', 'markdown'],
-          outputPath: `out-${counter}.json`,
-        }),
-        client,
-      )
-
-      const output = parseFormInstructions(post.mock.calls[0][1]).output as Record<string, unknown>
-      expect(output.formats).toEqual(['spatial', 'markdown'])
-      // Sending both keys is a 400 — assert absence explicitly, since a subset
-      // match would pass with `format` also present.
-      expect(output).not.toHaveProperty('format')
-    })
-
-    it('sends output.format (and never output.formats) for a single-format request', async () => {
-      const input = await writeInput()
-      const { client, post } = mockClient({ output: { markdown: '# Hello' } })
-
-      await performParseDocumentCall(extractArgs({ filePath: input, mode: 'text', format: 'markdown' }), client)
-
-      const output = parseFormInstructions(post.mock.calls[0][1]).output as Record<string, unknown>
-      expect(output.format).toBe('markdown')
-      expect(output).not.toHaveProperty('formats')
-    })
-
-    it('omits includeWords when unset, and never sends it on a markdown-only request', async () => {
-      const input = await writeInput()
-      const spatial = mockClient(spatialFixture)
-      const markdownOnly = mockClient({ output: { markdown: '# Hello' } })
-
-      await performParseDocumentCall(
-        extractArgs({ filePath: input, mode: 'structure', format: 'spatial', outputPath: `out-${counter}.json` }),
-        spatial.client,
-      )
-      await performParseDocumentCall(
-        extractArgs({ filePath: input, mode: 'text', format: 'markdown', includeWords: true }),
-        markdownOnly.client,
-      )
-
-      expect(parseFormInstructions(spatial.post.mock.calls[0][1]).output).not.toHaveProperty('includeWords')
-      expect(parseFormInstructions(markdownOnly.post.mock.calls[0][1]).output).not.toHaveProperty('includeWords')
-    })
-
-    it('writes both spatial and markdown to outputPath and mentions markdown bytes in the summary', async () => {
-      const input = await writeInput()
-      const outName = `out-${counter}.json`
-      const markdown = '# Quarterly Report'
-      const { client } = mockClient({ ...spatialFixture, output: { ...spatialFixture.output, markdown } })
-
-      const result = await performParseDocumentCall(
-        extractArgs({ filePath: input, mode: 'structure', formats: ['spatial', 'markdown'], outputPath: outName }),
-        client,
-      )
-
-      expect(result.isError).toBeFalsy()
-      const summary = text(result)
-      expect(summary).toContain('Extracted 3 elements')
-      expect(summary).toContain(`${Buffer.byteLength(markdown)} bytes of Markdown`)
-      expect(summary).toContain('output.markdown')
-      const written = await fs.promises.readFile(path.join(sandboxDir, outName), 'utf-8')
-      expect(JSON.parse(written).output.markdown).toBe(markdown)
-    })
-
-    it('requires outputPath for a multi-format request that includes spatial', async () => {
-      const input = await writeInput()
-      const { client, post } = mockClient(spatialFixture)
-
-      const result = await performParseDocumentCall(
-        extractArgs({ filePath: input, mode: 'structure', formats: ['spatial', 'markdown'] }),
-        client,
-      )
-
-      expect(result.isError).toBe(true)
-      expect(text(result)).toContain('outputPath')
-      expect(post).not.toHaveBeenCalled()
-    })
-
-    it('writes nothing when a multi-format response is missing the requested markdown', async () => {
-      const input = await writeInput()
-      const outName = `out-${counter}.json`
-      const { client } = mockClient(spatialFixture)
-
-      const result = await performParseDocumentCall(
-        extractArgs({ filePath: input, mode: 'structure', formats: ['spatial', 'markdown'], outputPath: outName }),
-        client,
-      )
-
-      expect(result.isError).toBe(true)
-      expect(text(result)).toContain('Nothing was written')
-      await expect(fs.promises.access(path.join(sandboxDir, outName))).rejects.toThrow()
     })
   })
 
-  describe('markdown-only formatting options', () => {
-    it('omits markdown-only options from the sent instructions when unset', async () => {
+  describe('schema / parseConfig / options nesting', () => {
+    it('nests schema directly under instructions', async () => {
       const input = await writeInput()
-      const { client, post } = mockClient({ output: { markdown: '# Hello' } })
+      const { client, post } = mockClient(extractFixture)
 
-      await performParseDocumentCall(extractArgs({ filePath: input, mode: 'text', format: 'markdown' }), client)
+      await performExtractFieldsCall(extractArgs({ filePath: input }), client)
 
       const instructions = parseFormInstructions(post.mock.calls[0][1])
-      const output = instructions.output as Record<string, unknown>
-      expect(output).not.toHaveProperty('useHtmlTables')
-      expect(output).not.toHaveProperty('enableSemanticBlockFormatting')
-      expect(output).not.toHaveProperty('includeHeadersAndFooters')
-      expect(output).not.toHaveProperty('extractWordsFromPictures')
+      expect(instructions.schema).toEqual(invoiceSchema)
     })
 
-    it('includes markdown-only options in the sent instructions when set', async () => {
+    it('nests language under parseConfig.options', async () => {
       const input = await writeInput()
-      const { client, post } = mockClient({ output: { markdown: '# Hello' } })
+      const { client, post } = mockClient(extractFixture)
 
-      await performParseDocumentCall(
-        extractArgs({
-          filePath: input,
-          mode: 'text',
-          format: 'markdown',
-          useHtmlTables: false,
-          enableSemanticBlockFormatting: false,
-          includeHeadersAndFooters: true,
-          extractWordsFromPictures: true,
-        }),
+      await performExtractFieldsCall(
+        extractArgs({ filePath: input, mode: 'structure', language: 'german' }),
         client,
       )
 
       const instructions = parseFormInstructions(post.mock.calls[0][1])
-      expect(instructions.output).toMatchObject({
-        useHtmlTables: false,
-        enableSemanticBlockFormatting: false,
-        includeHeadersAndFooters: true,
-        extractWordsFromPictures: true,
-      })
+      expect(instructions.parseConfig).toEqual({ mode: 'structure', options: { language: 'german' } })
     })
 
-    it('omits markdown-only options from the sent instructions when output is spatial-only', async () => {
+    it('omits parseConfig.options entirely when no language tuning is given', async () => {
       const input = await writeInput()
-      const { client, post } = mockClient(spatialFixture)
+      const { client, post } = mockClient(extractFixture)
 
-      await performParseDocumentCall(
-        extractArgs({
-          filePath: input,
-          mode: 'structure',
-          format: 'spatial',
-          outputPath: `out-${counter}.json`,
-          useHtmlTables: true,
-        }),
-        client,
-      )
+      await performExtractFieldsCall(extractArgs({ filePath: input }), client)
 
       const instructions = parseFormInstructions(post.mock.calls[0][1])
-      expect(instructions.output).not.toHaveProperty('useHtmlTables')
-    })
-  })
-
-  describe('language / maxLanguages / maxScripts', () => {
-    it('rejects maxLanguages in text mode, which does no OCR, rather than dropping it', async () => {
-      const input = await writeInput()
-      const { client, post } = mockClient({ output: { markdown: '# Hello' } })
-
-      const result = await performParseDocumentCall(
-        extractArgs({ filePath: input, mode: 'text', format: 'markdown', maxLanguages: 3 }),
-        client,
-      )
-
-      expect(result.isError).toBe(true)
-      expect(text(result)).toContain('text mode')
-      expect(post).not.toHaveBeenCalled()
+      expect(instructions.parseConfig).toEqual({ mode: 'understand' })
     })
 
-    it('sends auto-detection tuning under options on a happy path', async () => {
+    it('rejects maxLanguages when language is explicitly set, before any API call', async () => {
       const input = await writeInput()
-      const { client, post } = mockClient(spatialFixture)
+      const { client, post } = mockClient(extractFixture)
 
-      await performParseDocumentCall(
-        extractArgs({
-          filePath: input,
-          mode: 'structure',
-          format: 'spatial',
-          outputPath: `out-${counter}.json`,
-          maxLanguages: 3,
-          maxScripts: 1,
-        }),
-        client,
-      )
-
-      expect(parseFormInstructions(post.mock.calls[0][1]).options).toEqual({ maxLanguages: 3, maxScripts: 1 })
-    })
-
-    it('omits options entirely when no language tuning is given', async () => {
-      const input = await writeInput()
-      const { client, post } = mockClient(spatialFixture)
-
-      await performParseDocumentCall(
-        extractArgs({ filePath: input, mode: 'structure', format: 'spatial', outputPath: `out-${counter}.json` }),
-        client,
-      )
-
-      expect(parseFormInstructions(post.mock.calls[0][1])).not.toHaveProperty('options')
-    })
-
-    it('rejects maxLanguages when language is explicitly set', async () => {
-      const input = await writeInput()
-      const { client, post } = mockClient(spatialFixture)
-
-      const result = await performParseDocumentCall(
-        extractArgs({
-          filePath: input,
-          mode: 'structure',
-          format: 'spatial',
-          outputPath: `out-${counter}.json`,
-          language: 'german',
-          maxLanguages: 3,
-        }),
+      const result = await performExtractFieldsCall(
+        extractArgs({ filePath: input, language: 'german', maxLanguages: 3 }),
         client,
       )
 
@@ -627,68 +228,263 @@ describe('performParseDocumentCall', () => {
       expect(text(result)).toContain('maxLanguages')
       expect(post).not.toHaveBeenCalled()
     })
+
+    it('rejects maxScripts when language is explicitly set, before any API call', async () => {
+      const input = await writeInput()
+      const { client, post } = mockClient(extractFixture)
+
+      const result = await performExtractFieldsCall(
+        extractArgs({ filePath: input, language: 'german', maxScripts: 2 }),
+        client,
+      )
+
+      expect(result.isError).toBe(true)
+      expect(text(result)).toContain('maxScripts')
+      expect(post).not.toHaveBeenCalled()
+    })
+
+    it('omits includeCitations/strict/multimodal when unset', async () => {
+      const input = await writeInput()
+      const { client, post } = mockClient(extractFixture)
+
+      await performExtractFieldsCall(extractArgs({ filePath: input }), client)
+
+      const instructions = parseFormInstructions(post.mock.calls[0][1])
+      expect(instructions).not.toHaveProperty('options')
+    })
+
+    it('sends includeCitations/strict/multimodal under options when set', async () => {
+      const input = await writeInput()
+      const { client, post } = mockClient(extractFixture)
+
+      await performExtractFieldsCall(
+        extractArgs({ filePath: input, includeCitations: false, strict: true, multimodal: true }),
+        client,
+      )
+
+      const instructions = parseFormInstructions(post.mock.calls[0][1])
+      expect(instructions.options).toEqual({ includeCitations: false, strict: true, multimodal: true })
+    })
+
+    it('nests free-text instructions and storeRun alongside schema', async () => {
+      const input = await writeInput()
+      const { client, post } = mockClient(extractFixture)
+
+      await performExtractFieldsCall(
+        extractArgs({ filePath: input, instructions: 'Prefer the most recent invoice date.', storeRun: true }),
+        client,
+      )
+
+      const instructions = parseFormInstructions(post.mock.calls[0][1])
+      expect(instructions).toMatchObject({
+        instructions: 'Prefer the most recent invoice date.',
+        storeRun: true,
+      })
+    })
+  })
+
+  describe('response handling', () => {
+    it('returns output.data inline, pretty-printed', async () => {
+      const input = await writeInput()
+      const { client } = mockClient(extractFixture)
+
+      const result = await performExtractFieldsCall(extractArgs({ filePath: input }), client)
+
+      expect(result.isError).toBeFalsy()
+      const output = text(result)
+      expect(output).toContain(JSON.stringify(extractFixture.output.data, null, 2))
+    })
+
+    it('lists not_found field paths in the grounding signal, capped at 10 with a +N more suffix', async () => {
+      const input = await writeInput()
+      const manyNotFound: Record<string, unknown> = {}
+      const data: Record<string, unknown> = {}
+      for (let i = 0; i < 12; i++) {
+        manyNotFound[`field${i}`] = { match: 'not_found' }
+        data[`field${i}`] = null
+      }
+      const { client } = mockClient({ output: { data, metadata: manyNotFound, pages: [] } })
+
+      const result = await performExtractFieldsCall(extractArgs({ filePath: input }), client)
+
+      const notFoundLine = text(result)
+        .split('\n')
+        .find((line) => line.startsWith('Not found:'))
+      expect(notFoundLine).toContain('field0')
+      expect(notFoundLine).toContain('field9')
+      expect(notFoundLine).not.toContain('field10')
+      expect(notFoundLine).toContain('+2 more')
+    })
+
+    it('reports the citation match summary counts', async () => {
+      const input = await writeInput()
+      const { client } = mockClient(extractFixture)
+
+      const result = await performExtractFieldsCall(extractArgs({ filePath: input }), client)
+
+      const output = text(result)
+      expect(output).toContain('id_match: 1')
+      expect(output).toContain('fuzzy_match: 1')
+    })
+
+    it('writes the full response to outputPath and notes where citations landed, without leaking them inline', async () => {
+      const input = await writeInput()
+      const outName = `out-${counter}.json`
+      const { client } = mockClient(extractFixture)
+
+      const result = await performExtractFieldsCall(extractArgs({ filePath: input, outputPath: outName }), client)
+
+      expect(result.isError).toBeFalsy()
+      const output = text(result)
+      expect(output).toContain(outName)
+      expect(output).not.toContain(SECRET)
+      const written = await fs.promises.readFile(path.join(sandboxDir, outName), 'utf-8')
+      expect(written).toContain(SECRET)
+      expect(JSON.parse(written).output.metadata).toBeTruthy()
+    })
+
+    it('notes citations were omitted (not written) when outputPath is absent and metadata is non-empty', async () => {
+      const input = await writeInput()
+      const { client } = mockClient(extractFixture)
+
+      const result = await performExtractFieldsCall(extractArgs({ filePath: input }), client)
+
+      const output = text(result)
+      expect(output).toContain('omitted')
+      expect(output).toContain('outputPath')
+      expect(output).not.toContain(SECRET)
+    })
+
+    it('does not claim citations were returned when output.metadata is null', async () => {
+      const input = await writeInput()
+      const { client } = mockClient({ output: { data: { invoiceNumber: 'INV-001' }, metadata: null, pages: [] } })
+
+      const result = await performExtractFieldsCall(extractArgs({ filePath: input }), client)
+
+      expect(result.isError).toBeFalsy()
+      const output = text(result)
+      expect(output).not.toContain('citations were returned')
+      expect(output).not.toContain('Pass outputPath')
+    })
+
+    it('rejects a 2xx response whose output.data is not an object, writing nothing', async () => {
+      const input = await writeInput()
+      const outName = `out-${counter}.json`
+      const { client } = mockClient({ output: { data: 'not an object', metadata: {} } })
+
+      const result = await performExtractFieldsCall(extractArgs({ filePath: input, outputPath: outName }), client)
+
+      expect(result.isError).toBe(true)
+      expect(text(result)).toContain('output.data')
+      await expect(fs.promises.access(path.join(sandboxDir, outName))).rejects.toThrow()
+    })
+
+    it('surfaces runId and the credits/price_composition split in the success message', async () => {
+      const input = await writeInput()
+      const { client } = mockClient(extractFixture)
+
+      const result = await performExtractFieldsCall(extractArgs({ filePath: input }), client)
+
+      const output = text(result)
+      expect(output).toContain('run_abc123')
+      expect(output).toContain('10.5 Data Extraction credit')
+      expect(output).toContain('989.5 remaining')
+      expect(output).toContain('parse')
+      expect(output).toContain('extract')
+    })
+
+    it('sends the x-nutrient-api-version header', async () => {
+      const input = await writeInput()
+      const { client, post } = mockClient(extractFixture)
+
+      await performExtractFieldsCall(extractArgs({ filePath: input }), client)
+
+      expect(post).toHaveBeenCalledWith(expect.any(String), expect.anything(), {
+        'x-nutrient-api-version': '2026-05-25',
+      })
+    })
   })
 
   describe('API errors', () => {
-    it('explains a 402 as a credit balance problem rather than a transient failure', async () => {
+    it('renders a 402 as a non-retryable credit balance problem via handleExtractionApiError', async () => {
       const input = await writeInput()
       const { client } = mockErrorClient(402, {
         status: 402,
         requestId: 'req_402',
-        errorMessage: 'Insufficient credits. This request requires 2 credits, 0 remaining.',
+        errorMessage: 'Insufficient credits. This request requires 12 credits, 0 remaining.',
       })
 
-      const result = await performParseDocumentCall(
-        extractArgs({ filePath: input, mode: 'text', format: 'markdown' }),
-        client,
-      )
+      const result = await performExtractFieldsCall(extractArgs({ filePath: input }), client)
 
       expect(result.isError).toBe(true)
       const message = text(result)
       expect(message).toContain('Insufficient credits')
       expect(message).toContain('Retrying will not help')
       expect(message).toContain('req_402')
-    })
-
-    it('renders the Data Extraction error envelope, which the Processor handler does not recognize', async () => {
-      const input = await writeInput()
-      const { client } = mockErrorClient(400, {
-        status: 400,
-        requestId: 'req_400',
-        errorMessage: 'The request is malformed.',
-        errorDetails: {
-          source: 'request',
-          code: 'invalid_request',
-          failingPaths: [{ path: 'output.format', details: 'must be spatial or markdown' }],
-        },
-      })
-
-      const result = await performParseDocumentCall(
-        extractArgs({ filePath: input, mode: 'text', format: 'markdown' }),
-        client,
-      )
-
-      expect(result.isError).toBe(true)
-      const message = text(result)
-      expect(message).toContain('HTTP 400')
-      expect(message).toContain('The request is malformed.')
-      expect(message).toContain('invalid_request')
-      expect(message).toContain('output.format')
-      expect(message).not.toContain('Error processing API response')
+      // The advice must fit THIS endpoint: recommending text mode (which the
+      // parse tool offers) would send the caller into a guaranteed rejection.
+      expect(message).toContain('structure')
+      expect(message).not.toMatch(/\btext: 1 credit/)
     })
 
     it('surfaces a non-JSON error body with its status', async () => {
       const input = await writeInput()
       const { client } = mockErrorClient(503, 'upstream unavailable', { raw: true })
 
-      const result = await performParseDocumentCall(
-        extractArgs({ filePath: input, mode: 'text', format: 'markdown' }),
-        client,
-      )
+      const result = await performExtractFieldsCall(extractArgs({ filePath: input }), client)
 
       expect(result.isError).toBe(true)
       expect(text(result)).toContain('HTTP 503')
       expect(text(result)).toContain('upstream unavailable')
     })
+  })
+})
+
+// The handler tests above construct args directly, so they cannot catch a bad
+// default on the schema itself — parse real input here instead.
+describe('ExtractFieldsArgsSchema', () => {
+  const minimal = { filePath: 'in.pdf', schema: { type: 'object', properties: { total: { type: 'number' } } } }
+
+  it('leaves includeCitations, strict and multimodal undefined when unset', () => {
+    const parsed = ExtractFieldsArgsSchema.parse(minimal)
+
+    // A zod default here would ship an implicit `false` and silently disable
+    // citations, which the API enables by default, for every caller.
+    expect(parsed.includeCitations).toBeUndefined()
+    expect(parsed.strict).toBeUndefined()
+    expect(parsed.multimodal).toBeUndefined()
+  })
+
+  it('defaults mode to understand and storeRun to false', () => {
+    const parsed = ExtractFieldsArgsSchema.parse(minimal)
+
+    expect(parsed.mode).toBe('understand')
+    expect(parsed.storeRun).toBe(false)
+  })
+
+  it('rejects text mode, which this endpoint does not offer', () => {
+    expect(() => ExtractFieldsArgsSchema.parse({ ...minimal, mode: 'text' })).toThrow()
+  })
+
+  it('accepts every mode this endpoint does offer', () => {
+    for (const mode of ['structure', 'understand', 'agentic']) {
+      expect(ExtractFieldsArgsSchema.parse({ ...minimal, mode }).mode).toBe(mode)
+    }
+  })
+
+  it('requires a schema whose root is an object with properties', () => {
+    expect(() => ExtractFieldsArgsSchema.parse({ filePath: 'in.pdf' })).toThrow()
+    expect(() => ExtractFieldsArgsSchema.parse({ ...minimal, schema: { type: 'array', properties: {} } })).toThrow()
+  })
+
+  it('rejects free-text instructions beyond the documented 10000 characters', () => {
+    expect(() => ExtractFieldsArgsSchema.parse({ ...minimal, instructions: 'x'.repeat(10001) })).toThrow()
+    expect(
+      ExtractFieldsArgsSchema.parse({ ...minimal, instructions: 'x'.repeat(10000) }).instructions,
+    ).toHaveLength(10000)
+  })
+
+  it('rejects a url that is not a URL', () => {
+    expect(() => ExtractFieldsArgsSchema.parse({ schema: minimal.schema, url: 'not-a-url' })).toThrow()
   })
 })
